requirements.txt
streamlit
google-generativeai
pillow
pillow-heif